Refactor: Prepare before implement response API
